Update for FW 2.4. See changelog for details (#420)
* Drop support for Ubuntu 16.04, ROS kinetic
* Drop support for sensors on FW < 2.0
* Require C++ 14 for build
* Use sensor http interface for sensors on FW 2.1+
* Switch from nodes to nodelets in ouster_ros
* Separate ROS launch files for play, replay, and record
* Add new experimental ROS config services set_config and get_config
* Bugfixes. See Changelog

+++ ouster_viz/src/cloud.cpp
@@ -46,7 +46,7 @@ struct CloudIds {
 };
 
 bool GLCloud::initialized = false;
-GLfloat GLCloud::program_id;
+GLuint GLCloud::program_id;
 CloudIds GLCloud::cloud_ids;
 
 GLCloud::GLCloud(const Cloud& cloud) : point_size{cloud.point_size_} {
